feat:remove codemirror js/css in html
