+++ .eslintrc.js
@@ -10,7 +10,8 @@ module.exports = {
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'camelcase': 'off'
+    'camelcase': 'off',
+    'eqeqeq': 'off'
   },
   parserOptions: {
     parser: 'babel-eslint'
