+++ public/index.html
@@ -46,8 +46,8 @@
   <link rel="stylesheet" href="libs/css/code-themes/github-v2.min.css">
 
   <!-- codemirror -->
-  <link rel="stylesheet" href="libs/css/codemirror.min.css">
-  <link rel="stylesheet" href="libs/css/show-hint.css">
+  <!-- <link rel="stylesheet" href="libs/css/codemirror.min.css"> -->
+  <!-- <link rel="stylesheet" href="libs/css/show-hint.css"> -->
   <link rel="stylesheet" href="libs/css/style-mirror.css">
 
   <link rel="stylesheet" href="libs/css/animate.css">
@@ -59,41 +59,21 @@
 </head>
 
 <body>
-    <div class="loading" id="loading">
-    <div class="loading-wrapper">
-      <div class="loading-text">Loading...</div>
-      <div class="loading-anim"></div>
-    </div>
-  </div>
   <div id="app" >
 
   </div>
 
 
-  <!--应用主体-->
-
-  <script src="libs/scripts/marked.min.js"></script>
-
   <!-- codemirror -->
   <script src="libs/scripts/codemirror/codemirror.min.js"></script>
-  <script src="libs/scripts/codemirror/css.js"></script>
-  <script src="libs/scripts/codemirror/matchbrackets.js"></script>
-  <script src="libs/scripts/codemirror/active-line.js"></script>
-  <script src="libs/scripts/codemirror/show-hint.js"></script>
-  <script src="libs/scripts/codemirror/css-hint.js"></script>
-  <script src="libs/scripts/codemirror/format.js"></script>
 
   <script src="libs/scripts/markdown.min.js"></script>
   <script src="libs/scripts/prettify.min.js"></script>
-  <!-- <script src="libs/scripts/index.js"></script> -->
   <script src="libs/scripts/jquery.min.js"></script>
   <script src="libs/scripts/closebrackets.js"></script>
-
   <script src="assets/scripts/sync-scroll.js"></script>
-  <script src="assets/scripts/themes/default-theme.js"></script>
-  <script src="assets/scripts/util.js"></script>
+
   <script>
-    $('#loading').hide();
     window.console
       && window.console.log
       && (console.log("Think big, train fast, learn deep. See https://github.com/yanglbme"))
