+++ src/components/Loading.vue
@@ -1,0 +1,18 @@
+<template>
+    <div class="loading" id="loading">
+      <div class="loading-wrapper">
+        <div class="loading-text">Loading...</div>
+        <div class="loading-anim"></div>
+      </div>
+    </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style scoped>
+
+</style>
